Refactored upgrade-downgrade test, to be shared by different frameworks. Added it to Elastic tests.
