+++ frameworks/elastic/tests/test_shakedown.py
@@ -5,6 +5,7 @@ import sdk_install as install
 import sdk_tasks as tasks
 import sdk_marathon as marathon
 import sdk_utils as utils
+import sdk_test_upgrade
 
 DEFAULT_NUMBER_OF_SHARDS = 1
 DEFAULT_NUMBER_OF_REPLICAS = 1
@@ -126,6 +127,12 @@ def test_unchanged_scheduler_restarts_without_restarting_tasks():
 def test_kibana_proxylite_adminrouter_integration():
     # run this test as late as possible, as it takes 10+ minutes for kibana to be ready
     check_kibana_proxylite_adminrouter_integration()
+
+
+@pytest.mark.upgrade
+@pytest.mark.sanity
+def test_upgrade_downgrade():
+    sdk_test_upgrade.upgrade_downgrade(PACKAGE_NAME, DEFAULT_TASK_COUNT)
 
 
 @pytest.mark.recovery
