+++ frameworks/helloworld/tests/test_sanity.py
@@ -9,6 +9,7 @@ import sdk_install as install
 import sdk_marathon as marathon
 import sdk_spin as spin
 import sdk_tasks as tasks
+import sdk_test_upgrade
 
 from tests.config import (
     PACKAGE_NAME,
@@ -273,3 +274,9 @@ def test_lock():
     # Verify ZK is unchanged
     zk_config_new = shakedown.get_zk_node_data(zk_path)
     assert zk_config_old == zk_config_new
+
+
+@pytest.mark.upgrade
+@pytest.mark.sanity
+def test_upgrade_downgrade():
+    sdk_test_upgrade.upgrade_downgrade(PACKAGE_NAME, DEFAULT_TASK_COUNT)
